refactor(common): CHECKOUT-2951 Convert `config` module into TypeScript
